Update faq-topics.md
I removed mention of Object Store and Substrate - these are internal names. 
I also reworded how we use search crawler as it sounded like we were using it to trigger the AI clustering, which we're not. We use it to identify changes.
I made a few adjustments to replace generate with discover as we don't want to confuse users that we're creating knowledge - we just discover from what is already in content. For example, the difference between saying the topic is "Fruit" when we are discovering Apples and Oranges in the data source, vs. identifying "Apples" and "Oranges" as the topics.