Updated with the new Card Designer section
Added a new section ##Add a Card Designer  to demonstrate the designer.
In the available cards, Microsoft has introduced a **Card Designer** to add the card and customize the design and look and feel of the card. Also the quick view layout can be editted and modified using the existing adaptivecards samples or a new design can be created and updated in the cards.