Update add-viva-connections-app.md
I discussed this with @pamgreen-msft on a call today. I'd guess the majority of people setting up Viva Connections want to pin the app to the top of the Teams App Bar. That's how they have seen it shown in all of the demos and screenshots from Microsoft. When I was setting up Viva Connections for a client today, I sort of knew I needed to pin the app, but I slid right past the buttle about it in this article.

So, my suggested edit is to include the "fastest path to joy" to accomplish this, with a link to info about Setup Policies rather than just linking to that article, which covers many other aspects of the topic.