Update closing-and-scoring.md
https://microsoft-ce-csi.acrolinx.cloud/api/v1/checking/scorecards/ce17925c-1cab-46ab-bd36-51e37efcf75b#CORRECTNESS

Line 193: ...you might want to close the objective with the progress you're it at that given point in time. > ...you might want to close the objective with the progress you've made at that given point in time. 

(I changed the wording slightly because "it at" was unclear.)

**Suggestion:** Revise the sentence for concision and clarity:
This is because substantial has been made, and you might want to close the objective.