Update configuring-okrs.md
configuring-okrs.md
https://microsoft-ce-csi.acrolinx.cloud/api/v1/checking/scorecards/08e036ed-1a25-4151-8a98-10d460fc31c8#CORRECTNESS
Lines 69 & 71: drop down > dropdown list
Line 89: metric based > metric-based
Line 195: engg. > eng.
Line 232: The check-in owner will be able to check-in on the OKR (manual check-in) and set up a data link on the OKR (to auto-check-in).  > The check-in owner will be able to check in on the OKR (manual check-in) and set up a data link on the OKR (to auto check in). 
Line 240: drop-down > dropdown list