Update configure-successfactors-content-source.md
Line 145: Change image code to custom Markdown to try to indent the image. 
Line 176: The troubleshooting comment for this step is a direct response to the mention of an error within the step. Delete the Troubleshooting heading. However, @bhaswatic, review the subsequent text to confirm it is in the right section: Admin center configuration.