Rename "Duration" metadata column as "ContentDuration" to prevent internal field name collisions
SharePoint already has a site column with internal name "Duration" - this is field with GUID {4D54445D-1C84-4a6d-B8DB-A51DED4E1ACC} which is used in calendar event schema and is a hidden field of type integer.

I strongly advise and recommend development team adopts a new less generic name such as "ContentDuration" so that a Viva Learning content type may be created at the Content Type Hub level and pushed down through syndication. If we keep "Duration" as the internal site column name, we are welcoming unnecessary trouble in future iterations should Microsoft decide to make the field global. Please reconsider the internal name of this field which will require documentation and code updates in the Viva Learning service.