Update insight-faqs.md
Line 32: One or more of the employees on your team are not assigned a Viva Insights license. > One or more of the employees on your team isn't assigned a Viva Insights license.
Line 33: Or the latest available HR data uploaded into the advanced insights app did not include the latest hierarchy or employee data. > Or the latest available HR data uploaded into the advanced insights app didn't include the latest hierarchy or employee data.
Line 37: Currently, you cannot configure settings for these insights. > Currently, you can't configure settings for these insights.
Line 41: If the team structure does not meet the minimum group size, the team's average is not used or shown in the metrics. > If the team structure doesn't meet the minimum group size, the team's average isn't used or shown in the metrics.
Line 45: While this capability is currently unavailable, it is planned for a future release.  > While this capability is currently unavailable, it's planned for a future release.
Line 51: They do not have Viva Insights enabled. > They don't have Viva Insights enabled.
Line53 : The team does not meet the minimum group size limit. > The team doesn't meet the minimum group size limit.