Update privacy.md
Was receiving a suggestion on this submission "•	Line 218, Column 2:  [Suggestion] Column headers are required for tables. Add appropriate header text, or remove any extra table cells that cause an extra column."

Table was not missing headers and had no extra, but had a lot of extra space. Deleted extra space and resubmitting.