Update power-bi-integration.md
Fixing the image indents under the numbers.

I added lightboxes since I am unable to get new images- does that not meet the requirement in the comment?
"...If cropping these images or obtaining new images isn't possible or practical, try lightbox coding so readers can zoom on large images."

Line 56- sentence is not a part of the above step. Neither is the image below it so I did not change them.