Update meeting-query.md
Lines 82-90: Add pipes to the ends of each row. Not following the accepted Markdown format is blocking.
Line 96: metrics metrics > metrics
Line 120: Add quotes around the operator and value content.
Line 127: Add space between the end of a sentence and the beginning of the next sentence. 
Lines 140-155: Test content in Markdown columns

Line 120: Is "Recurring" supposed to be capitalized?