Update edit-viva-home.md
IT admins are asking for a way to retrieve the current state of the default landing to the home site before making a change. IT admins always do not have Viva Connections enabled as some orgs have IT managed by partners . Proposing a change to retrieve the current state of default landing to the home site before making a change.