Apply suggestions from code review
Line 25: Change ellipses to emdash. 
Line 33: Delete "learn:" to create "In this article, you'll (- verb something. - verb something else.)
Lines 35-38. Add periods.
Line 41: will become > becomes
Line 62: Add a period to the end of the sentence.
Lines 84 and 90: Add a space before the hyphen to match the rest of the list items.
Line 110: Delete a space in the link text and add a period to the end of the sentence.