Updated errant localization metadata
Everything in Viva/solutions/viva-goals-healthy-okr-program needed its metadata field updated to the correct field/value other than "kr-types.md". As you can see, the trend (or perhaps universal status: I didn't check too closely) was "localization_priority: Priority". Because "kr-types.md", the only one with the correct field, was designated as "medium", I defined the rest as "medium" as well. While it's possible "Priority" meant "high" in the eyes of the original writer, I went with the safer "medium", as it is the default, does not require human translation, will come online more quickly, and matches the only correctly formatted file in the folder. Plus, we can always change machine translation to human translation, but we can't reverse the man hours it would take to do things the other way around.