Correct form and target of link
This link was changed before the build system started checking links that looked like site-relative links (i.e., start with a forward slash). The link also included the file name extension, which should never be on a site-relative link, and it pointed to the very short article that this link is in. Look at file history, I see that this link once pointed to complex-organization-insights, which is now in officearchive-pr.