Update understanding-views.md
PM said it was fine apart from 3 minor typos. I corrected two of the typos. The other, which involved the "to log in to Viva Goals, visit" link not opening a new window, I did my best to fix by removing the naked link. However, I'm not sure this is something that can be fixed from our end. In any case, at least we're not getting dinged for a style error anymore. Also, the freshness date has obviously been updated.