Global Admin Changes Part 1
Removed "clear no" references to global admins. I will point out here that there was a single line item in the Excel doc, specifically in reference to roles-permissions-in-viva-goals, that actually contained multiple GA references, one of fell into the "In the absence of X" category that was elsewhere classified as a "Yes". As such, I made only the change to that line that was not part of the "In the absence of X" category, leaving that to require approval from an outside source.