Added adjusting notifications info
Support has been getting an uptick in support tickets from customers asking how to turn off certain notifications. This info was in the FAQ at the end of the document, but I've moved it up and made it a H3 section so it will appear in the "In this article" at the top.