Incorporate web link article with create dashboard
The Web link card article is no longer needed. Took the intro from that article and added it to the Web Link card section in the create-dashboard.md article.

Also added SharePoint and Sway to the URL/SSO table at the bottom of the document.

Deleted use-the-link-card.md article

Updated toc and redirect json file