Update employee-retention-metric.md
Removed spreadsheet image of data collection timeline because, 1) the dates don't match the example, 2) the text was small and hard to see, 3) the expanded view couldn't show the entire image (too wide to fit in the screen). The image also had a fourth column that is post-collection but could be confusing. Also the lightbox handle blocked the text rendering it useless.